# src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: Prova Azienda
  version: 1.0.0
paths:
...
 /workTime: 
  post:
      summary: Add new WorkTime 
      description: Add a new WorkTime
      operationId: addWorkTime
      requestBody:
        description: Create a new Worktime
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkTime'
          application/xml:
            schema:
              $ref: '#/components/schemas/WorkTime'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/WorkTime'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkTime'          
            application/xml:
              schema:
                $ref: '#/components/schemas/WorkTime'
        '405':
          description: Invalid input 
 /workTime/{workTimeId}:   
  get:
      summary: Find WorkTime by ID
      description: Returns a single WorkTime
      operationId: getWorkTimeById
      parameters:
        - name: workTimeId
          in: path
          description: ID of workTime to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkTime'          
            application/xml:
              schema:
                $ref: '#/components/schemas/WorkTime'
        '400':
          description: Invalid ID supplied
        '404':
          description: WorkTime not found
  post:
      summary: Updates a workTime  
      description: ''
      operationId: updateWorkTime
      parameters:
        - name: workTimeId
          in: path
          description: ID of workTime that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
        - name: StartTime
          in: query
          description: StartTime of project that needs to be updated
          schema:
            type: string
        - name: EndTime
          in: query
          description: EndTime of project that needs to be updated
          schema:
            type: string
        - name: ApprovedDate
          in: query
          description: ApprovedDate of project that needs to be updated
          schema:
            type: string
            
      responses:
        '405':
          description: Invalid input        
  delete:
      summary: Deletes a WorkTime
      description: delete a workTime
      operationId: deleteWorkTime
      parameters:
        - name: api_key
          in: header
          description: ''
          required: false
          schema:
            type: string
        - name: workTimeId
          in: path
          description: WorkTime id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid workTime value      
components:
  schemas:
...
    WorkTime:
      type: object
      properties:
       id:
          type: integer
          format: int64
       projectId:
          type: integer
          format: int64
       employeeId:
          type: integer
          format: int64
       startTime:
           type: string
           format: date-time      
       endTime:
           type: string
           format: date-time
       approvedBy:
           type: string
       approvedDate:
           type: string
           description: Start date
           example: "2017-01-01"
           format: date
           pattern: "YYYY-MM-DD"
           minLength: 0
           maxLength: 10